[FLINK-4512] [FLIP-10] Add option to persist periodic checkpoints
[FLINK-4509] [FLIP-10] Specify savepoint directory per savepoint
[FLINK-4507] [FLIP-10] Deprecate savepoint backend config

This closes #2608.

+++ flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/ZooKeeperCheckpointIDCounter.java
@@ -23,6 +23,7 @@ import org.apache.curator.framework.recipes.shared.SharedCount;
 import org.apache.curator.framework.recipes.shared.VersionedValue;
 import org.apache.curator.framework.state.ConnectionState;
 import org.apache.curator.framework.state.ConnectionStateListener;
+import org.apache.flink.runtime.jobgraph.JobStatus;
 import org.apache.flink.runtime.jobmanager.HighAvailabilityMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -91,30 +92,17 @@ public class ZooKeeperCheckpointIDCounter implements CheckpointIDCounter {
 	}
 
 	@Override
-	public void shutdown() throws Exception {
+	public void shutdown(JobStatus jobStatus) throws Exception {
 		synchronized (startStopLock) {
 			if (isStarted) {
 				LOG.info("Shutting down.");
 				sharedCount.close();
 				client.getConnectionStateListenable().removeListener(connStateListener);
 
-				LOG.info("Removing {} from ZooKeeper", counterPath);
-				client.delete().deletingChildrenIfNeeded().inBackground().forPath(counterPath);
-
-				isStarted = false;
-			}
-		}
-	}
-
-	@Override
-	public void suspend() throws Exception {
-		synchronized (startStopLock) {
-			if (isStarted) {
-				LOG.info("Suspending.");
-				sharedCount.close();
-				client.getConnectionStateListenable().removeListener(connStateListener);
-
-				// Don't remove any state
+				if (jobStatus.isGloballyTerminalState()) {
+					LOG.info("Removing {} from ZooKeeper", counterPath);
+					client.delete().deletingChildrenIfNeeded().inBackground().forPath(counterPath);
+				}
 
 				isStarted = false;
 			}
